fixed comments review №2 (#10)
* Убрал ошибку в тестах ПОСТМАН

* Убрал ошибку с кодировкой при сборке проекта

* Убрал ошибку в тестах ПОСТМАН

* Убрал ошибку в тестах ПОСТМАН

* Убрал ошибку в тестах ПОСТМАН

* Функциональность «Лента событий».

* Обновил

* fix: corrected FilmLikesDaoImpl and FilmDaoImplTest after final api-tests

* change director methods name
all tests done

* change 4 methods

* fix: скорректировал методы dao по замечаниям ревью

* fix: по замечаниям ревью скорректировал методы сервиса, переименовал классы, исправил метод getAllReviews, добавил фиг.скобки в условия

* change ids types

* change ids types

* fix: corrected README.md accordance with new features

* fix: по замечаниям ревью убрал неуказанные в ТЗ ограничения на размер столбцов таблиц, изменил в таблицах тип столбца id на Long,
скорректировал findRecommendationsFilms

* fix: по замечаниям ревью исправил имя столбцов на id в таблицах: directors, reviews, events

* fix: по замечаниям ревью исправил readme и схему БД

* fix: по замечаниям ревью исправил readme и схему БД

* fix: по замечаниям ревью исправил readme и схему БД

---------

+++ src/main/java/ru/yandex/practicum/filmorate/storage/dao/film/FilmLikesDao.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate.storage.dao.film;
 
 public interface FilmLikesDao {
-    void like(Integer id, Integer userId);
+    void like(Long id, Long userId);
 
-    void unlike(Integer id, Integer userId);
+    void unlike(Long id, Long userId);
 }
